requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
requests>=2.26.0
nest-asyncio>=1.5.1
quart>=0.18.0
hypercorn>=0.14.0 